requirements.txt
Flask
requests
rembg
Pillow
gunicorn
